Added the most important file
This notebook contains all the analysis, visualization, and the LSTM model. It helps you to understand the data well and the process of how to go about working with datasets in general. I have plotted different kinds of plots, loss, and accuracy plots. I have also done an exploratory data analysis.